Create MSYS2.yaml and fix some issues on MSYS2.
Add MSYS2 building action to check building issues on MSYS2.

Fix some "error=char-subscripts" for isdigit and isspace functions
by adding explicit cast to int,
which is required by C99 standard and later.

+++ src/console/dlt-receive-v2.c
@@ -188,7 +188,7 @@ int64_t convert_arg_to_byte_size(char *arg)
 
     /* check if valid input */
     for (i = 0; i < strlen(arg) - 1; ++i)
-        if (!isdigit(arg[i]))
+        if (!isdigit((int)arg[i]))
             return -2;
 
     /* last character */
@@ -200,7 +200,7 @@ int64_t convert_arg_to_byte_size(char *arg)
         factor = 1024 * 1024;
     else if ((arg[strlen(arg) - 1] == 'G') || (arg[strlen(arg) - 1] == 'g'))
         factor = 1024 * 1024 * 1024;
-    else if (!isdigit(arg[strlen(arg) - 1]))
+    else if (!isdigit((int)arg[strlen(arg) - 1]))
         return -2;
 
     /* range checking */
